Issue #883: Extract tenant_routing into its own component
Moves the pure tenant-schema resolution logic out of mdr_auth into a
new top-level component so other auth-gated LIF services (Advisor, LIF
API) can reuse the same sanitize/derive/resolve contract without taking
a dependency on MDR. Addresses cbeach47's review note on PR #902:

  > Is this the time to at least take the tenant schema logic and
  > place it in the general components auth folder (with the intent
  > to reuse for Advisor, LIF API, etc)?

Pure rename + import updates. No behavior change.

- `components/lif/mdr_auth/tenant.py` → `components/lif/tenant_routing/core.py`
  (via `git mv`, so history follows).
- `components/lif/tenant_routing/__init__.py` re-exports the public
  surface so callers can `from lif.tenant_routing import ...`.
- 3 import sites updated: the MDR auth middleware, the MDR tenant
  service, and the test file.
- `test/components/lif/mdr_auth/test_tenant.py` moves to
  `test/components/lif/tenant_routing/test_core.py` to stay co-located
  with its subject.
- `projects/lif_mdr_api/pyproject.toml` gains the new brick under
  `[tool.polylith.bricks]`.

No other projects referenced `lif.mdr_auth.tenant`, so no other
pyproject.toml updates needed. When Advisor/LIF API adopt tenant
routing, they'll add the same brick line.
